agentic-llm/nat_wrapper/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[tool.setuptools_scm]
git_describe_command = "git describe --long --first-parent"
root = "../.."

[project]
name = "nat_5g_slicing"
dynamic = ["version"]
dependencies = [
    "nvidia-nat[langchain]",
    "gpudb>=7.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "langgraph>=0.2.0",
    "langchain-nvidia-ai-endpoints>=0.3.0",
    "psutil>=5.9.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "arize-phoenix>=3.0.0"
]
requires-python = ">=3.11,<3.14"
description = "5G Network Slicing Agentic Workflow with NeMo Agent Toolkit"
keywords = ["ai", "agents", "5g", "network-slicing"]
classifiers = ["Programming Language :: Python"]

[project.entry-points.'nat.components']
nat_5g_slicing = "nat_5g_slicing.register"
